% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.tdcm}
\alias{data.tdcm}
\alias{data.tdcm01}
\alias{data.tdcm02}
\alias{data.tdcm03}
\alias{data.tdcm04}
\title{Several data sets for the \code{tdcm} package.}
\format{
\code{data.tdcm01} is simulated sample data that has two time points, four
attributes, twenty items, one group of size 1000, and a single Q-matrix. The
format is a list of two:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{qmatrix}: a data frame specifying the Q-matrix
}

\code{data.tdcm02} is simulated data that has three time points, two
attributes, ten items, one group of size 2500, and a single Q-matrix. The
format is a list of two:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{qmatrix}: a data frame specifying the Q-matrix
}

\code{data.tdcm03} is simulated data that has three time points, two
attributes, one group of size 1500, and three different ten-item Q-matrices
for each time point. Anchor items are specified as items 1/1/21 and items
14/24. The format is a list of five:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{q1}: a data frame specifying the Q-matrix for the first time point
\item \code{q2}: a data frame specifying the Q-matrix for the second time point
\item \code{q3}: a data frame specifying the Q-matrix for the third time point
\item \code{qmatrix123}: data frame specifying the combined Q-matrix for all time
points
}

\code{data.tdcm04} is simulated data that has two time points, four
attributes, twenty items, two group of size 800 and 900, respectively, and a
single Q-matrix. The format is a list of three:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{qmatrix}: a data frame specifying the Q-matrix
\item \code{groups}: a vector specifying the examinee group memberships
}
}
\usage{
data.tdcm01

data.tdcm02

data.tdcm03

data.tdcm04
}
\description{
Several data sets for the \code{tdcm} package.
}
\keyword{data}
