% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcdm.R
\name{hcdm}
\alias{hcdm}
\title{Estimation of the Hierarchical Diagnostic Classification Model (HDCM)}
\usage{
hcdm(data, q.matrix, hierarchy, linkfct = "logit", ...)
}
\arguments{
\item{data}{A required \eqn{N \times T \times I} \code{matrix} or
\code{data.frame} where rows correspond to \code{N} examinees and columns
represent
the binary item responses across \code{T} time points and \code{I} items.}

\item{q.matrix}{A required \eqn{I \times A} \code{matrix} indicating
which items measure which attributes. If there are multiple
Q-matrices,
then they must have the same number of attributes and must be
stacked on top of each other for estimation (to specify multiple
Q-matrices, see \code{num.q.matrix}, \code{num.items}, and \code{anchor}).}

\item{hierarchy}{A string containing restrictions of the hierarchy.
For example, \code{'Attr1 > Attr2 > Attr3'} means that attribute 3 requires
attribute 2, which requires attribute 1.}

\item{linkfct}{A \code{string} or a \code{vector} indicating the LCDM
link function. Currently accepts "logit" (default) to estimate
the LCDM, "identity" to estimate the GDINA model, and "log" link
function to estimate the reduced reparameterized unified model
(RRUM).
The link function vector must have length equal to the total
number of items across all time points.}

\item{...}{Extra arguments to pass to the \code{\link[CDM:gdina]{CDM::gdina()}} call.}
}
\value{
An object of class \code{hcdm} which inherits from \link[CDM:gdina]{CDM::gdina}
}
\description{
\code{hcdm()} estimates the hierarchical diagnostic classification model
(HCDM; Templin & Bradshaw, 2025).
}
\references{
Templin J, Bradshaw L. (2014). Hierarchical Diagnostic Classification Models:
A Family of Models for Estimating and Testing Attribute Hierarchies.
\emph{Psychometrika}, 79_(2), 317-339. \url{doi:10.1007/s11336-013-9362-0}.
}
