% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg.summary.param1.R
\name{mg.summary.param1}
\alias{mg.summary.param1}
\title{Utility function in \pkg{TDCM}}
\usage{
mg.summary.param1(model, time.points, num.atts, num.items)
}
\arguments{
\item{model}{gdina object from mg.tdcm function}

\item{time.points}{number of time points}

\item{num.atts}{number of attributes}

\item{num.items}{number of items}
}
\description{
Utility function in \pkg{TDCM}
}
\keyword{internal}
