% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg.summary.param2.R
\name{mg.summary.param2}
\alias{mg.summary.param2}
\title{Utility function in \pkg{TDCM}}
\usage{
mg.summary.param2(model, time.points, num.atts, num.items)
}
\arguments{
\item{model}{gdina object from mg.tdcm function}

\item{time.points}{number of time points}

\item{num.atts}{number of attributes}

\item{num.items}{number of items}
}
\description{
Utility function in \pkg{TDCM}
}
