% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg.summary2.R
\name{mg.summary2}
\alias{mg.summary2}
\title{Utility function in \pkg{TDCM}}
\usage{
mg.summary2(
  model,
  num.atts,
  num.groups,
  time.points,
  attribute.names,
  group.names
)
}
\arguments{
\item{model}{gdina object from mg.tdcm function}

\item{num.atts}{number of attributes}

\item{num.groups}{number of groups}

\item{time.points}{number of time points}

\item{attribute.names}{optional vector to specify attribute names}

\item{group.names}{optional vector to specify group names}
}
\description{
Utility function in \pkg{TDCM}
}
\keyword{internal}
