% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.option2.R
\name{summary.option2}
\alias{summary.option2}
\title{Utility function in \pkg{TDCM}}
\usage{
\method{summary}{option2}(model, num.atts, time.points, attribute.names)
}
\arguments{
\item{model}{gdina object from tdcm function}

\item{num.atts}{number of attributes}

\item{time.points}{number of time points}

\item{attribute.names}{optional argument to specify attribute names}
}
\description{
Utility function in \pkg{TDCM}
}
