% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.param.R
\name{summary.param}
\alias{summary.param}
\title{Utility function in 'tdcm'.}
\usage{
\method{summary}{param}(model, num.atts, numitems, time.points)
}
\arguments{
\item{model}{model from \bold{tdcm} estimation}

\item{num.atts}{number of attributes}

\item{numitems}{number of items at each time point}

\item{time.points}{number of time points}
}
\description{
Utility function in 'tdcm'.
}
\keyword{internal}
