% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm-package.R
\docType{package}
\name{TDCM-package}
\alias{TDCM}
\alias{TDCM-package}
\title{The Transition Diagnostic Classification Model Framework}
\description{
The \emph{R} package \pkg{TDCM} enables longitudinal diagnostic classification model (DCM) analyses
within the transition diagnostic classification model (TDCM) framework.
}
\details{
Diagnostic classification models (DCMs) are psychometric models designed to classify
examinees according to their proficiency on a set of categorical latent traits, known as
attributes. Longitudinal DCMs have been developed as psychometric options for modeling changes
in attribute proficiency over time.

The \pkg{TDCM} implements estimation of the transition DCM (TDCM; Madison & Bradshaw, 2018a),
a longitudinal extension of the log-linear cognitive diagnosis model (LCDM; Henson, Templin, &
Willse, 2009). As the LCDM subsumes many other DCMs, many other DCMs can be estimated
longitudinally via the TDCM. The package includes functions to estimate the single-group and
multigroup TDCM, summarize results of interest including item parameters, growth proportions,
transition probabilities, transitional reliability, attribute correlations, model fit, and
growth plots.

For more details and examples, see \code{vignette("TDCM", package = "TDCM")}.
}
\references{
de la Torre, J. (2011). The Generalized DINA model framework. \emph{Psychometrika, 76}, 179–199.
\url{doi:10.1007/s11336-011-9207-7}.

George, A. C., Robitzsch, A., Kiefer, T., Gross, J., & Ünlü , A. (2016). The \emph{R} package \strong{CDM}
for cognitive diagnosis models. \emph{Journal of Statistical Software,74}(2), 1-24.
\url{doi:10.18637/jss.v074.i02}

Henson, R., Templin, J., & Willse, J. (2009). Defining a family of cognitive diagnosis models
using log-linear models with latent variables. \emph{Psychometrika, 74}, 191-21.
\url{doi:10.1007/s11336-008-9089-5}.

Johnson, M. S., & Sinharay, S. (2020). The reliability of the posterior probability of skill
attainment in diagnostic classification models. \emph{Journal of Educational Measurement, 47}(1),
5–31. \url{doi:10.3102/1076998619864550}.

Kaya, Y., & Leite, W. (2017). Assessing change in latent skills across time With longitudinal
cognitive diagnosis modeling: An evaluation of model performance.
\emph{Educational and Psychological Measurement, 77}(3), 369–388.
\url{doi:10.1177/0013164416659314}.

Li, F., Cohen, A., Bottge, B., & Templin, J. (2015). A latent transition analysis model for
assessing change in cognitive skills. \emph{Educational and Psychological Measurement, 76}(2),
181–204.
\url{doi:10.1177/0013164415588946}.

Madison, M. J. (2019). Reliably assessing growth with longitudinal diagnostic classification
models. \emph{Educational Measurement: Issues and Practice, 38}(2), 68-78.
\url{doi:10.1111/emip.12243}.

Madison, M. J., & Bradshaw, L. (2018a). Assessing growth in a diagnostic classification model
framework. \emph{Psychometrika, 83}(4), 963-990.
\url{doi:10.1007/s11336-018-9638-5}.

Madison, M. J., & Bradshaw, L. (2018b). Evaluating intervention effects in a diagnostic
classification model framework. \emph{Journal of Educational Measurement, 55}(1), 32-51.
\url{doi:10.1111/jedm.12162}.

Madison, M.J., Chung, S., Kim, J., & Bradshaw, L.P. (2024) Approaches to estimating longitudinal
diagnostic classification models. \emph{Behaviormetrika, 51}(7), 7-19.
\url{doi:10.1007/s41237-023-00202-5}.

Madison, M. J., Jeon, M., Cotterell, M., Haab, S., & Zor, S. (2025). TDCM: An R package for estimating
longitudinal diagnostic classification models. \emph{Multivariate Behavioral Research}, 1–10.
\url{doi:10.1080/00273171.2025.2453454}.

Ravand, H., Effatpanah, F., Kunina-Habenicht, O., & Madison, M. J. (2025). A didactic illustration
of writing skill growth through a longitudinal diagnostic classification model. \emph{Frontiers in Psychology}.
\url{doi:10.3389/fpsyg.2024.1521808}.

Rupp, A. A., Templin, J., & Henson, R. (2010).
\emph{Diagnostic Measurement: Theory, Methods, and Applications}. New York: Guilford.
ISBN: 9781606235430.

Schellman, M., & Madison, M. J. (2024). Estimating the reliability of skill transition in longitudinal DCMs.
\emph{Journal of Educational and Behavioral Statistics}.\url{doi:10.3102/10769986241256032}.

Templin, J., & Bradshaw, L. (2013). Measuring the reliability of diagnostic classification model
examinee estimates. \emph{Journal of Classification, 30}, 251-275.
\url{doi:10.1007/s00357-013-9129-4}.

Wang. S., Yang. Y., Culpepper, S. A., & Douglas, J. (2018). Tracking skill acquisition with
cognitive diagnosis models: A higher-order, hidden Markov model With covariates.
\emph{Journal of Educational and Behavioral Statistics, 43}(1), 57-87.
\url{doi:10.3102/1076998617719727}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cotterell.github.io/tdcm/}
  \item \url{https://github.com/cotterell/tdcm}
  \item Report bugs at \url{https://github.com/cotterell/tdcm/issues}
}

}
\author{
The \pkg{TDCM} package was created by:
\itemize{
\item Matthew J. Madison, University of Georgia, mjmadison@uga.edu
\item Sergio Haab, University of Iowa, sergio-haab@uiowa.edu
\item Constanza Mardones-Segovia, University of Georgia, constanza.mardonessegovia@uga.edu
\item Minjeong Jeon, University of California - Los Angeles, mjjeon@ucla.edu
\item Michael E. Cotterell, University of Georgia, mepcott@uga.edu
}

More information about the \pkg{TDCM} package authors, copyright holders, funders, etc. is
available in the package's \code{DESCRIPTION} file. You can see this information by executing
\code{utils::packageDescription("TDCM")}.
}
\keyword{internal}
