% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.base.R
\name{tdcm.base}
\alias{tdcm.base}
\title{Utility function for base estimation in TDCMs.}
\usage{
tdcm.base(data, qnew, dcmrule)
}
\arguments{
\item{data}{item response data}

\item{qnew}{stacked Q-matrix}

\item{dcmrule}{specific DCM to estimate}
}
\description{
Utility function for base estimation in TDCMs.
}
