% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.ipd.R
\name{tdcm.ipd}
\alias{tdcm.ipd}
\title{Assessing item parameter drift (IPD) in the Transition Diagnostic Classification Model (TDCM)}
\usage{
tdcm.ipd(model)
}
\arguments{
\item{model}{A \code{tdcm} object returned from the \code{\link{tdcm}} function.}
}
\value{
A list with the following items:
\itemize{
\item \code{$ipd.stats}: Data frame containing results of item-wise Wald tests.

\item \code{$coef}: Data frame containing item parameter estimates for each time point.

\item \code{$estimates}: List of \eqn{\lambda} vectors containing all item parameter estimates.

\item \code{$item.probs.time}: List with predicted item response probabilities for each time point.

}
}
\description{
The \code{tdcm.ipd()} function assesses item parameter drift (IPD) in the TDCM (e.g., Madison & Bradshaw, 2018) by applying the Wald test
for differential item functioning (de la Torre, 2011; Hou, de la Torre & Nandakumar, 2014).
The \emph{p}-values are also calculated by a Holm adjustment for multiple comparisons. In the case of two time
points, an effect size of item parameter drift (labeled as UA in the ipd.stats value)
is defined as the weighted absolute difference of item response functions.
}
\examples{
\donttest{

############################################################################
# Example 1: TDCM with full measurement invariance
############################################################################

# Load dataset: T=2, A=4
data(data.tdcm01, package = "TDCM")
data <- data.tdcm01$data
q.matrix <- data.tdcm01$q.matrix
# Estimate model
model1 <- TDCM::tdcm(data, q.matrix, num.time.points = 2, invariance = TRUE,
                    rule = "LCDM", num.q.matrix = 1)

# Run IPD analysis
ipd = tdcm.ipd(model1)
ipd$ipd.stats
ipd$coef
ipd$parameters

}
}
\references{
de la Torre, J. (2011). The Generalized DINA model framework. \emph{Psychometrika, 76}, 179–199.
\url{doi:10.1007/s11336-011-9207-7}.

Hou, L., de la Torre, J., & Nandakumar, R. (2014). Differential item functioning assessment
in cognitive diagnostic modeling: Application of the Wald test to investigate DIF
in the DINA model. \emph{Journal of Educational Measurement, 51}, 98-125. \url{doi:10.1111/jedm.12036}.

Madison, M. J., & Bradshaw, L. (2018a). Assessing growth in a diagnostic classification model
framework. \emph{Psychometrika}, \strong{83}(4), 963-990.
\url{doi:10.1007/s11336-018-9638-5}.
}
