% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.mq.R
\name{tdcm.mq}
\alias{tdcm.mq}
\title{Utility function to estimate TDCM with multiple Q-matrices.}
\usage{
tdcm.mq(
  data,
  qmatrix,
  time.points,
  invariance = TRUE,
  dcmrule = "GDINA",
  number.q = 1,
  num.items = c(),
  anchor = c()
)
}
\arguments{
\item{data}{item response data}

\item{qmatrix}{specified qumatrices}

\item{time.points}{number of time points}

\item{invariance}{invariance assumption (T or F)}

\item{dcmrule}{specific DCM to estimate}

\item{number.q}{number of Q-matrices}

\item{num.items}{number of items for each Q-matrix}

\item{anchor}{anchor items specified as pairs in a vector}
}
\description{
Utility function to estimate TDCM with multiple Q-matrices.
}
\keyword{internal}
