% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.plot.mg.R
\name{tdcm.plot.mg}
\alias{tdcm.plot.mg}
\title{Utility function in \pkg{TDCM}}
\usage{
tdcm.plot.mg(results, attribute.names = c(), group.names = c())
}
\arguments{
\item{results}{results from \bold{mg.tdcm.summary}}

\item{attribute.names}{optional vector of attribute names to include in plots.}

\item{group.names}{optional vector of group names to include in plots.}
}
\description{
Utility function in \pkg{TDCM}
}
\keyword{internal}
