% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.rel.R
\name{tdcm.rel}
\alias{tdcm.rel}
\title{Utility function to compute transition reliability.}
\usage{
tdcm.rel(model, numatts, time.points, transition.option, attribute.names = c())
}
\arguments{
\item{model}{gdina object from tdcm estimation}

\item{numatts}{number of attributes}

\item{time.points}{number of time points}

\item{transition.option}{transition.option specified in summary function}

\item{attribute.names}{optional attribute names}
}
\value{
Several reliability metrics
}
\description{
Utility function to compute transition reliability.
}
\keyword{internal}
