% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.rule.R
\docType{data}
\name{tdcm.rules}
\alias{tdcm.rules}
\title{TDCM Condensation Rules}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\usage{
tdcm.rules
}
\description{
A \emph{condensation rule} is a formula that states how different attributes
combine to form an observed or latent response (Rupp, Templin, & Henson,
2010). The \pkg{TDCM} package includes support for \code{"LCDM"}, \code{"DINA"},
\code{"DINO"}, \code{"CRUM"}, \code{"RRUM"}, \code{"LCDM1"} for the LCDM with only main effects,
\code{"LCDM2"} for the LCDM with two-way interactions, \code{"LCDM3"}, and so on.
Evaluate \code{TDCM::tdcm.rules$TDCM} for a complete list of condensation rules
supported by the \pkg{TDCM} package.
}
\examples{
TDCM::tdcm.rules$TDCM

}
\references{
Rupp, A. A., Templin, J., & Henson, R. (2010).
\emph{Diagnostic Measurement: Theory, Methods, and Applications}. New York:
Guilford. ISBN: 9781606235430.
}
\keyword{datasets}
