% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tdcm_check_dims}
\alias{tdcm_check_dims}
\title{Check the number of time points for TDCM data.}
\usage{
tdcm_check_dims(ir_data, q_data)
}
\arguments{
\item{ir_data}{An \code{ir_data} parameter supplied to \code{tdcm()}.}

\item{q_data}{An \code{q_data} parameter supplied to \code{tdcm()}.}
}
\description{
Check the number of time points for TDCM data.
}
\keyword{internal}
