% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.tdcm}
\alias{data.tdcm}
\alias{data.tdcm01}
\alias{data.tdcm02}
\alias{data.tdcm03}
\alias{data.tdcm04}
\alias{data.tdcm05}
\title{Several data sets for the \pkg{TDCM} package.}
\format{
\code{data.tdcm01} is a simulated dataset with two time points, four
attributes, twenty items, one group of size 1000, and a single Q-matrix. The
format is a list of two:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{q.matrix}: a data frame specifying the Q-matrix
}

\code{data.tdcm02} is a simulated dataset with three time points, two
attributes, ten items, one group of size 2500, and a single Q-matrix. The
format is a list of two:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{q.matrix}: a data frame specifying the Q-matrix
}

\code{data.tdcm03} is a simulated dataset with three time points, two
attributes, one group of size 1500, and three different ten-item Q-matrices
for each time point. Anchor items are specified as items 1/11/21 and items
14/24. The format is a list of five:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{q.matrix.1}: a data frame specifying the Q-matrix for the first time point
\item \code{q.matrix.2}: a data frame specifying the Q-matrix for the second time point
\item \code{q.matrix.3}: a data frame specifying the Q-matrix for the third time point
\item \code{q.matrix.stacked}: data frame specifying the combined Q-matrix for all time
points
}

\code{data.tdcm04} is a simulated dataset with two time points, four
attributes, twenty items, two group of size 800 and 900, respectively, and a
single Q-matrix. The format is a list of three:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{q.matrix}: a data frame specifying the Q-matrix
\item \code{groups}: a vector specifying the examinee group membership
}

\code{data.tdcm05} is a simulated dataset with one time point, four
attributes, and twenty items. For use with the 1-PLCDM. The format is a list
of two:
\itemize{
\item \code{data}: a data frame of binary item responses
\item \code{q.matrix}: a data frame specifying the Q-matrix
}
}
\usage{
data.tdcm01

data.tdcm02

data.tdcm03

data.tdcm04

data.tdcm05
}
\description{
Several data sets for the \pkg{TDCM} package.
}
\examples{
\donttest{

#############################
## Example 1: T = 2, A = 4 ##
data(data.tdcm01, package = "TDCM")
data <- data.tdcm01$data
q.matrix <- data.tdcm01$q.matrix
model <- TDCM::tdcm(data, q.matrix, num.time.points = 2)
results <- TDCM::tdcm.summary(model)
results$item.parameters
results$growth.effects

#############################
## Example 3: T = 3, A = 2 ##
data <- data.tdcm03$data
q1 <- data.tdcm03$q.matrix.1
q2 <- data.tdcm03$q.matrix.2
q3 <- data.tdcm03$q.matrix.3
q <- data.tdcm03$q.matrix.stacked

#TDCM with anchor items constrained
m1 <- tdcm(data, q, num.time.points = 3, num.q.matrix = 3,
anchor = c(1,11,1,21,14,24), num.items = c(10,10,10))

#TDCM without anchor items
m2 <- tdcm(data, q, num.time.points = 3, num.q.matrix = 3, num.items = c(10,10,10))

#Compare models to assess measurement invariance
tdcm.compare(m1, m2)

#Summarize model 1
r1 = tdcm.summary(m1, transition.option = 3)
r1$item.parameters
r1$growth
r1$growth.effects

}
}
\keyword{data}
