% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm-package.R
\docType{package}
\name{tdcm-package}
\alias{tdcm-package}
\title{tdcm: Transition Diagnostic Classification Models}
\description{
The **tdcm** package implements estimation of the transition DCM (TDCM; Madison & Bradshaw, 2018), a longitudinal extension of the log-linear cognitive diagnosis model (LCDM; Henson, Templin, & Willse, 2009). As the LCDM subsumes many other DCMs, many other DCMs can be estimated longitudinally via the TDCM. The package includes functions to estimate the single-group and multigroup TDCM, summarize results of interest including item parameters, growth proportions, transition probabilities, transitional reliability, attribute correlations, model fit, and growth plots.
}
\details{
See \code{vignette("tdcm", package = "tdcm")} or for an overview of the \code{tdcm}
package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cotterell/tdcm}
  \item Report bugs at \url{https://github.com/cotterell/tdcm/issues}
}

}
\author{
\strong{Maintainer}: Michael E. Cotterell \email{mepcott@uga.edu} (\href{https://orcid.org/0000-0002-4302-0891}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Matthew Madison \email{mjmadison@uga.edu} (\href{https://orcid.org/0000-0002-2944-7442}{ORCID}) [copyright holder]
  \item Sergio Haab \email{sergio.haab@uga.edu}
  \item Minjeong Jeon \email{mjjeon@ucla.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item University of Georgia (The content and opinions expressed in this material do not necessarily reflect the views of nor are they endorsed by the University of Georgia or the University System of Georgia.) [copyright holder]
  \item Institute of Education Sciences (This work is supported by the U.S. Department of Education Institute of Education Sciences under IES Award Number R305D220020.) [funder]
  \item National Science Foundation (This work is supported by the National Science Foundation under NSF Award Number 1921373.) [funder]
}

}
\keyword{internal}
