% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm-package.R
\docType{package}
\name{tdcm-package}
\alias{tdcm-package}
\title{tdcm: Transition Diagnostic Classification Models}
\description{
Estimate the transition diagnostic classification model (TDCM) described in Madison & Bradshaw (2018) \doi{https://doi.org/10.1007/s11336-018-9638-5}, a longitudinal extension of the log-linear cognitive diagnosis model (LDCM) in Henson, Templin & Willse (2009) \doi{10.1007/s11336-008-9089-5}. As the LCDM subsumes many other diagnoistic classificaton models (DCMs), many other DCMs can be estimated longitudinally via the TDCM. The 'tdcm' package includes functions to estimate the single-group and multigroup TDCM, summarize results of interest including item parameters, growth proportions, transition probabilities, transitional reliability, attribute correlations, model fit, and growth plots.
}
\details{
See \code{vignette("tdcm", package = "tdcm")} or for an overview of the 'tdcm'
package.
}
\references{
\itemize{
\item de la Torre, J. (2011). The generalized DINA model framework.
Psychometrika, 76, 179-199.
\item George, A. C., Robitzsch, A., Kiefer, T., Gross, J., & Ünlü , A. (2016).
The R package CDM for cognitive diagnosis models. Journal of Statistical
Software, 74(2), 1-24.
\item Henson, R., Templin, J., & Willse, J. (2009). Defining a family of
cognitive diagnosis models using log linear models with latent variables.
Psychometrika, 74, 191-21.
\item Johnson, M. S., & Sinharay, S. (2020). The reliability of the posterior
probability of skill attainment in diagnostic classification models.
Journal of Educational Measurement, 47(1), 5 – 31.
\item Kaya, Y., & Leite, W. (2017). Assessing change in latent skills across time
with longitudinal cognitive diagnosis modeling: An evaluation of model
performance. Educational and Psychological Measurement, 77(3), 369–388.
\item Li, F., Cohen, A., Bottge, B., & Templin, J. (2015). A latent transition
analysis model for assessing change in cognitive skills. Educational and
Psychological Measurement, 76(2), 181–204.
\item Madison, M. J. (2019). Reliably assessing growth with longitudinal
diagnostic classification models. Educational Measurement: Issues and
Practice, 38(2), 68-78.
\item Madison, M. J., & Bradshaw, L. (2018a). Assessing growth in a diagnostic
classification model framework. Psychometrika, 82(4), 963-990.
\item Madison, M. J., & Bradshaw, L. (2018b). Evaluating intervention effects in
a diagnostic classification model framework. Journal of Educational
Measurement, 55(1), 32-51.
\item Madison, M.J., Chung, S., Kim, J. et al. Approaches to estimating
longitudinal diagnostic classification models. Behaviormetrika (2023).
\item Rupp, A. A., Templin, J., & Henson, R. (2010). Diagnostic measurement:
Theory, methods, and applications. New York: Guilford.
\item Schellman, M., & Madison, M. J. (2021, July). Estimating the reliability of
skill transition in longitudinal DCMs. Paper presented at the 2021
International Meeting of the Psychometric Society.
\item Templin, J., & Bradshaw, L. (2013). Measuring the reliability of diagnostic
classification model examinee estimates. Journal of Classification, 30,
251-275.
\item Wang. S., Yang. Y., Culpepper, S. A., & Douglas, J. (2018). Tracking Skill
Acquisition with cognitive diagnosis models: A higher-order, hidden Markov
model with covariates. Journal of  Educational and Behavioral Statistics,
43(1), 57-87.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cotterell/tdcm}
  \item Report bugs at \url{https://github.com/cotterell/tdcm/issues}
}

}
\author{
\strong{Maintainer}: Michael E. Cotterell \email{mepcott@uga.edu} (\href{https://orcid.org/0000-0002-4302-0891}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Matthew Madison \email{mjmadison@uga.edu} (\href{https://orcid.org/0000-0002-2944-7442}{ORCID}) [copyright holder]
  \item Sergio Haab \email{sergio.haab@uga.edu}
  \item Minjeong Jeon \email{mjjeon@ucla.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item University of Georgia (The content and opinions expressed in this material do not necessarily reflect the views of nor are they endorsed by the University of Georgia or the University System of Georgia.) [copyright holder]
  \item Institute of Education Sciences (This work is supported by the U.S. Department of Education Institute of Education Sciences under IES Award Number R305D220020.) [funder]
  \item National Science Foundation (This work is supported by the National Science Foundation under NSF Award Number 1921373.) [funder]
}

}
\keyword{internal}
