% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.R
\name{tdcm}
\alias{tdcm}
\title{Fitting Transition Diagnostic Classification Models (TDCMs)}
\usage{
tdcm(ir_data, q_data, invariance = TRUE, dcm_rule = "GDINA")
}
\arguments{
\item{ir_data}{The item response data, a required \eqn{T \times N \times I}
\code{array} where \eqn{T} is the number of time points, \eqn{N} is the number of
responses, and \eqn{I} is the number of items. The \code{ir_from()} function can
be used to help prepare this argument using values from a two-dimensional
\code{array}, \code{data.frame}, or \code{matrix}.}

\item{q_data}{The Q-matrix data, a required \eqn{T \times I \times A} \code{array}
where \eqn{T} is the number of time points, \eqn{I} is the number of items,
and \eqn{A} is the number of attributes. The \code{qmatrix_from()} function can be
used to help prepare this argument using values from a two-dimensional
\code{array}, \code{data.frame}, or \code{matrix}.}

\item{invariance}{A \code{logical} (i.e., boolean) value that indicates whether
item parameter invariance is constrained to be equal at each time point. The
default value is \code{TRUE}. If set to \code{FALSE}, item parameters are not assumed
to be equal over time.}

\item{dcm_rule}{A string or vector of condensation rules. Refer to the \code{rule}
parameter of \code{CDM::gdina} for more information. The default value is
\code{"GDINA"}.  If \code{dcm_rule} is a string, the condensation rule applies to all
items. If \code{dcm_rule} is a vector, then the condesnation rules are applied
item-wise.}
}
\value{
An object of class \code{tdcm} with entries as indicated in the CDM
package. For the TDCM-specific results (e.g., growth, transitions), results
are summarized using the tdcm.summary() function.
}
\description{
The \code{tdcm} Function calibrates and fits the transition
diagnostic classification model (TDCM; Madison & Bradshaw, 2018a), a
longitudinal extension of the log-linear cognitive diagnosis model (LCDM;
Henson, Templin, & Willse, 2009). To calibrate a TDCM involving multiple
groups, refer to the \code{mgtdcm} function.
}
\details{
Estimation is performed using the \code{CDM} package according to the approach
described in Madison et al. (2023).
}
\references{
\itemize{
\item de la Torre, J. (2011). The generalized DINA model framework.
\emph{Psychometrika, 76}, 179-199.
\item George, A. C., Robitzsch, A., Kiefer, T., Gross, J., & Ünlü , A. (2016).
The R package CDM for cognitive diagnosis models. \emph{Journal of Statistical
Software, 74}(2), 1-24.
\item Henson, R., Templin, J., & Willse, J. (2009). Defining a family of cognitive
diagnosis models using log linear models with latent variables.
\emph{Psychometrika, 74}, 191-21.
\item Madison, M. J., & Bradshaw, L. (2018a). Assessing growth in a diagnostic
classification model framework. \emph{Psychometrika, 82}(4), 963-990.
\item Madison, M. J., & Bradshaw, L. (2018b). Evaluating intervention effects in
a diagnostic classification model framework. \emph{Journal of Educational
Measurement, 55}(1), 32-51.
\item Madison, M.J., Chung, S., Kim, J., & Bradshaw, L. (2023). Approaches to
estimating longitudinal diagnostic classification models.
\emph{Behaviormetrika}.
\item Rupp, A. A., Templin, J., & Henson, R. (2010).
\emph{Diagnostic measurement: Theory, methods, and applications}.
New York: Guilford.
}
}
