% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdcm.R
\name{tdcm}
\alias{tdcm}
\title{Estimating the Transition Diagnostic Classification Model (TDCM)}
\usage{
tdcm(
  data,
  q.matrix,
  num.time.points,
  invariance = TRUE,
  rule = "GDINA",
  num.q.matrix = 1,
  num.items = c(),
  anchor = c(),
  progress = FALSE
)
}
\arguments{
\item{data}{A required \eqn{N \times T \times I} data matrix containing binary item responses.
For each time point, the binary item responses are in the columns.}

\item{q.matrix}{A required \eqn{I \times A} matrix indicating which items measure which
attributes. If there are multiple Q-matrices, then they must have the same number of attributes
and must be stacked on top of each other for estimation (to specify multiple Q-matrices, see
\code{num.q.matrix}, \code{num.items}, and \code{anchor}).}

\item{num.time.points}{A required integer \eqn{\ge 2} specifying the number of time points (i.e.,
measurement / testing occasions).}

\item{invariance}{logical. If \code{TRUE} (the default), the item parameter invariance will be
constrained to be equal at each time point. If \code{FALSE}, item parameters are not assumed to be
equal over time.}

\item{rule}{A string or a vector of the specific DCM to be employed. Currently accepts the
same values as \code{rule} in \code{\link[CDM:gdina]{CDM::gdina()}}. The default is \code{"GDINA"}, which is implemented with a
logit link to estimate the LCDM. If \code{rule} is supplied as a single string, then that DCM will
be assumed for each item. If entered as a vector, a DCM can be specified for each item.}

\item{num.q.matrix}{An optional integer specifying the number of Q-matrices. For many
applications, the same assessment is administered at each time point and this number is 1 (the
default). If there are different Q-matrices for each time point, then this argument must be
specified and should be equal to the number of time points. For example, if there are three time
points, and the Q-matrix for each time point is different, then \code{num.q.matrix = 3}. If there are
three time points, and the Q-matrix is different only for time point 3, then \code{num.q.matrix} is
still specified as \code{3}.}

\item{num.items}{An optional integer specifying the number of Q-matrices (the default is \code{1}).
when there are multiple Q-matrices, the number of items in each Q-matrix is specified as a
vector of length \code{T}. For example, if there are three time points, and the Q-matrices for each
time point have 8, 10, and 12 items, respectively, then \code{num.items = c(8, 10, 12)}. Default is an
empty vector to indicate there is only one Q-matrix.}

\item{anchor}{When there are different tests at each time point, this optional anchor argument is
a vector of pairs of item numbers indicating which items are the same across time points and
should be held invariant. For example, if there are three Q-matrices with 10 items each, and
Items 1, 11, and 21 are the same, and Items 14 and 24 are the same, then
\code{anchor = c(1,11,1,21,14,24)}. Default is an empty vector to indicate there is only one
Q-matrix.}

\item{progress}{logical. If \code{FALSE} (the default), the function will print the progress of
estimation. If \code{TRUE}, no progress information is printed.}
}
\value{
An object of class \code{gdina} with entries as described in \code{\link[CDM:gdina]{CDM::gdina()}}. To see a
TDCM-specific summary of the object (e.g., growth, transitions), use \code{\link[=tdcm.summary]{tdcm.summary()}}.
}
\description{
\code{tdcm()} is used to estimate the transition diagnostic classification model (TDCM; Madison &
Bradshaw, 2018a), which is a longitudinal extension of the log-linear cognitive diagnosis model
(LCDM; Henson, Templin, & Willse, 2009). It allows for the specification of many specific DCMs
via the \code{rule} option. For the multigroup TDCM, see \code{\link[=mg.tdcm]{mg.tdcm()}}.
}
\details{
Estimation of the TDCM via the \pkg{CDM} package (George, et al., 2016), which is based
on an EM algorithm as described in de la Torre (2011). The estimation approach is further
detailed in Madison et al. (2023).
}
\examples{
\donttest{
## Example 1: T = 2, A = 4
data(data.tdcm01, package = "TDCM")
data <- data.tdcm01$data
q.matrix <- data.tdcm01$q.matrix

# Estimate full TDCM with invariance assumed.
model1 <- TDCM::tdcm(data, q.matrix, num.time.points = 2)

# Summarize results with tdcm.summary().
results <- TDCM::tdcm.summary(model1, num.time.points = 2)
results$item.parameters
results$growth
results$transition.probabilities
}
}
\references{
de la Torre, J. (2011). The Generalized DINA Model Framework. \emph{Psychometrika} \strong{76}, 179–199.
\url{doi:10.1007/s11336-011-9207-7}

George, A. C., Robitzsch, A., Kiefer, T., Gross, J., & Ünlü , A. (2016). The \emph{R} package \strong{CDM}
for Cognitive Diagnosis Models. \emph{Journal of Statistical Software}, \strong{74}(2), 1-24.
\url{doi:10.18637/jss.v074.i02}

Henson, R., Templin, J., & Willse, J. (2009). Defining a Family of Cognitive Diagnosis Models
Using Log-Linear Models with Latent Variables. \emph{Psychometrika}, \strong{74}, 191-21.
\url{doi:10.1007/s11336-008-9089-5}

Johnson, M. S., & Sinharay, S. (2020). The Reliability of the Posterior Probability of Skill
Attainment in Diagnostic Classification Models. \emph{Journal of Educational Measurement}, \strong{47}(1),
5–31. \url{doi:10.3102/1076998619864550}

Kaya, Y., & Leite, W. (2017). Assessing Change in Latent Skills Across Time With Longitudinal
Cognitive Diagnosis Modeling: An Evaluation of Model Performance.
\emph{Educational and Psychological Measurement}, \strong{77}(3), 369–388.
\url{doi:10.1177/0013164416659314}

Li, F., Cohen, A., Bottge, B., & Templin, J. (2015). A Latent Transition Analysis Model for
Assessing Change in Cognitive Skills. \emph{Educational and Psychological Measurement}, \strong{76}(2),
181–204.
\url{doi:10.1177/0013164415588946}

Madison, M. J. (2019). Reliably Assessing Growth with Longitudinal Diagnostic Classification
Models. \emph{Educational Measurement: Issues and Practice}, \strong{38}(2), 68-78.
\url{doi:10.1111/emip.12243}

Madison, M. J., & Bradshaw, L. (2018a). Assessing Growth in a Diagnostic Classification Model
Framework. \emph{Psychometrika}, \strong{83}(4), 963-990.
\url{doi:10.1007/s11336-018-9638-5}

Madison, M. J., & Bradshaw, L. (2018b). Evaluating Intervention Effects in a Diagnostic
Classification Model Framework. \emph{Journal of Educational Measurement}, \strong{55}(1), 32-51.
\url{doi:10.1111/jedm.12162}

Madison, M.J., Chung, S., Kim, J., & Bradshaw, L.P. (2024) Approaches to estimating longitudinal
diagnostic classification models. \emph{Behaviormetrika}, \strong{51}(7), 7-19.
\url{doi:10.1007/s41237-023-00202-5}

Rupp, A. A., Templin, J., & Henson, R. (2010).
\emph{Diagnostic Measurement: Theory, Methods, and Applications}. New York: Guilford.
ISBN: 9781606235430.

Schellman, M., & Madison, M. J. (2021, July 20-23). Estimating the Reliability of Skill
Transition in Longitudinal DCMs. Paper presented virtually during a "Live Stage" session at the
\emph{2021 International Meeting of the Psychometric Society} (IMPS 2021).
\href{https://www.psychometricsociety.org/sites/main/files/file-attachments/imps_2021_virtual_schedule.pdf}{pdf:imps_2021_virtual_schedule.pdf}

Templin, J., & Bradshaw, L. (2013). Measuring the Reliability of Diagnostic Classification Model
Examinee Estimates. \emph{Journal of Classification}, \strong{30}, 251-275.
\url{doi:10.1007/s00357-013-9129-4}

Wang. S., Yang. Y., Culpepper, S. A., & Douglas, J. (2018). Tracking Skill Acquisition With
Cognitive Diagnosis Models: A Higher-Order, Hidden Markov Model With Covariates.
\emph{Journal of Educational and Behavioral Statistics}, \strong{43}(1), 57-87.
\url{doi:10.3102/1076998617719727}
}
